services/notification/requirements.txt
Flask
flask-socketio
kafka-python
python-dotenv
eventlet
